refactor: refactor components, extract styles
